Remove clients package. Move CLIs in airctrl
